refactor blacklist to global package
