+++ pkg/utils/global/global.go
@@ -1,4 +1,4 @@
-package internal
+package global
 
 import v1 "k8s.io/api/apps/v1"
